# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel V12 Patient and Appointment Management API - FSJ26",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/appointments/filter": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointments in a date range",
                "operationId": "55d543aafbc56f2a30cba3056176d68f",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Fecha de inicio (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-01-10"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Fecha de fin (opcional)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-02-15"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de citas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/appointments/patient": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Obtener pacientes asignados al doctor logueado",
                "operationId": "3a79d1a56e8e926455502706afb0f14a",
                "responses": {
                    "200": {
                        "description": "Lista de citas con datos del paciente y del doctor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "date_appointment": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2025-04-25"
                                            },
                                            "patient": {
                                                "type": "string",
                                                "example": "Ana Gómez"
                                            },
                                            "doctor": {
                                                "type": "string",
                                                "example": "Dr. Carlos Martínez"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado si no es doctor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access denied, only doctors can view this information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "log in to the system",
                "operationId": "e369fa054810db7ae55ecc66684da441",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "string",
                                            "example": "Juan Perez"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJK..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Logs out the authenticated user",
                "operationId": "c213530478940c4cde326564b658f385",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensaje": {
                                            "type": "string",
                                            "example": "Se ha cerrado la sesion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token returned at login",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}